+++ RELEASE_NOTES.md
@@ -1,3 +1,15 @@
+### 1.16.0 - 13.05.2015
+* Simplify ofObservableBuffered and toBlockingSeq
+
+### 1.15.0 - 30.03.2015
+* Add AsyncSeq.getIterator (unblocks use of AsyncSeq in FSharpx.Async)
+
+### 1.14 - 30.03.2015
+* Cancellable AsyncSeq.toBlockingSeq
+* Fix AsyncSeq.scanAsync to also return first state
+* Add a signature file
+* AsyncSeq got extracted as separate project and is now a dependency - https://github.com/fsprojects/FSharpx.Async/pull/24 
+
 ### 1.13 - 27.03.2015
 * Renamed to FSharp.Control.AsyncSeq
 * Remove surface area